added new function cloud-sync
new functionality to sync the local files with the cloud files. Before, cloud-sync,
this step had to be done outside of the cdectl.

function read-conf-file now checks for cloudname mismatch

when called for actions that require a cloudname, read-conf-file checks to ensure
the name provided is the same as in the configuration file. This protects somewhat
for modifying the layout outside ofe cdectl.